backend/requirements.txt
Flask
Flask-CORS
ultralytics
opencv-python-headless
Pillow
numpy
torch
torchvision